распределил имеющийся код по модулям
